packages/HDF5_BLS/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "HDF5_BLS"
# Use alpha release (1.0.2a0) for testing purposes, then beta release (1.0.2b0) for stable pre-release, release candidate (1.0.2rc0) for release candidate, and final release (1.0.2) for stable release
# Use major.minor.patch format for releases, with the patch version incremented for each release
version = "1.1.0"
description = "A package to unify the storage of Brillouin Light Scattering related data in HDF5 files"
readme = { file = "README.md", content-type = "text/markdown" }
license = "GPL-3.0-or-later"
authors = [
  { name = "Pierre Bouvet", email = "pierre.bouvet@meduniwien.ac.at" }
]
requires-python = ">=3.6"
keywords = ["HDF5", "Brillouin", "BLS"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "h5py>=3.12",
  "numpy>=2.0",
  "Pillow>=10.0.1",
  "scipy>=1.10",
  "sif_parser>=0.3",
  "openpyxl",
  "pyperclip",
  "pandas>=2.0.3",
  "brimfile",
  "matplotlib>=3.7.1",
]

[project.urls]
"Homepage" = "https://github.com/bio-brillouin/HDF5_BLS"
"Documentation" = "https://hdf5-bls.readthedocs.io/en/latest/"
"Issues" = "https://github.com/bio-brillouin/HDF5_BLS/issues"

[project.optional-dependencies]
dev = [
  "pytest",
]
docs = [
  "sphinx",
  "sphinx-rtd-theme",
]

# --- SETUPLOOLS CONFIGURATION FOR SRC LAYOUT ---

[tool.setuptools]
include-package-data = true
packages = ["HDF5_BLS", "HDF5_BLS.load_formats", "HDF5_BLS.brimfile_converter"]

# This maps the package name "HDF5_BLS" to the "src" directory relative to this pyproject.toml. This is CRUCIAL for the 'src' layout.
[tool.setuptools.package-dir]
"HDF5_BLS" = "src/_HDF5_BLS"
"HDF5_BLS.load_formats" = "src/_HDF5_BLS/load_formats"
"HDF5_BLS.brimfile_converter" = "src/_HDF5_BLS/brimfile_converter"


